Data is now being pushed from tab 2 to tab 3. Detailed view for processes is now complete. Some minor details and aesthetics things remain before the application is ready for additional features.

+++ 3DPrintMeApp/process.h
@@ -13,7 +13,7 @@
 //Different attributes are defined here. As this class only holds data, no methods are required to be defined.
 @property (strong,nonatomic) NSString *name;
 @property float cost;
-@property int resolution;
+@property float resolution;
 @property (strong,nonatomic) NSString *quality;
 @property (strong,nonatomic) NSString *quickinfo;
 @property (strong,nonatomic) NSString *materialusage;
